+++ src/utils/distance.js
@@ -1,3 +1,6 @@
 export default (u, v) => {
-  return Math.hypot(u.x - v.x, u.y - v.y);
+  return Math.hypot(
+    u.x - v.x, 
+    u.y - v.y
+  );
 };